requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
cffi==1.17.0
distlib==0.3.8
Django>=3.2
django-cors-headers>=4.4.0
djangorestframework<4.0  
djongo==1.3.6
dnspython==2.6.1
filelock==3.15.4
h11==0.14.0
idna==3.8
outcome==1.3.0.post0
platformdirs==4.2.2
psycopg2==2.9.9
pycparser==2.22
pymongo==3.12.0
PySocks==1.7.1
selenium==4.23.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.2.4
trio==0.26.2
pytz==2023.3
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
virtualenvwrapper-win==1.2.7
websocket-client==1.8.0
wsproto==1.2.0
gunicorn==20.1.0  # or the latest version
Pillow==10.4.0